Move session logic into sessions.go
